+++ groups/urls.py
@@ -1,11 +1,17 @@
 """The URLconf of the groups app."""
 
+from django.contrib.sitemaps.views import sitemap
 from django.urls import path
 
-from . import feeds, views
+from . import feeds, sitemaps, views
 
 #: The URL namespace of the groups app.
 app_name = 'groups'
+
+_sitemaps = {
+    'template_name': 'image-sitemap.xml',
+    'sitemaps': {'groups': sitemaps.GroupSitemap}
+}
 
 #: The URL patterns of the groups app.
 urlpatterns = [
@@ -13,6 +19,7 @@ urlpatterns = [
     path('<int:g_id>/', views.group, name='group'),
     path('<int:g_id>.atom', feeds.GroupAtom(), name='group.atom'),
     path('<int:g_id>.rss', feeds.GroupRSS(), name='group.rss'),
+    path('<section>-sitemap.xml', sitemap, _sitemaps, name='sitemap.xml'),
 ]
 
 __all__ = ['app_name', 'urlpatterns']